organize tasks into separate dir
