--- app/worker/__init__.py
@@ -1,3 +1,0 @@
-'''
-Module for tasks to be sent on task queue
-'''
